binding fix and code optimixation for documentList service
