+++ src/main/java/com/strandls/document/service/DocumentService.java
@@ -19,13 +19,9 @@ import com.strandls.document.pojo.BibTexItemType;
 import com.strandls.document.pojo.BulkUploadExcelData;
 import com.strandls.document.pojo.DocumentCreateData;
 import com.strandls.document.pojo.DocumentEditData;
-import com.strandls.document.pojo.DocumentListData;
 import com.strandls.document.pojo.DocumentUserPermission;
 import com.strandls.document.pojo.DownloadLogData;
 import com.strandls.document.pojo.ShowDocument;
-import com.strandls.esmodule.pojo.MapSearchParams;
-import com.strandls.esmodule.pojo.MapSearchQuery;
-import com.strandls.document.pojo.MapAggregationResponse;
 import com.strandls.taxonomy.pojo.SpeciesGroup;
 import com.strandls.user.pojo.Follow;
 import com.strandls.userGroup.pojo.Featured;
@@ -102,5 +98,7 @@ public interface DocumentService {
 	public List<Language> getLanguages(Boolean isDirty);
 
 	public Boolean documentDownloadLog(HttpServletRequest request, DownloadLogData downloadLogData);
+	
+	public void produceToRabbitMQ(String documentId, String updateType);
 
 }
