+++ src/main/java/com/strandls/document/service/Impl/DocumentServiceModule.java
@@ -5,6 +5,7 @@ package com.strandls.document.service.Impl;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
+import com.strandls.document.service.DocumentListService;
 import com.strandls.document.service.DocumentService;
 
 /**
@@ -14,9 +15,9 @@ import com.strandls.document.service.DocumentService;
 public class DocumentServiceModule extends AbstractModule {
 
 	@Override
-	protected void configure() {
+	protected void configure() 	{
 		bind(DocumentService.class).to(DocumentServiceImpl.class).in(Scopes.SINGLETON);
-		bind(DocumentListServiceImpl.class).in(Scopes.SINGLETON);
+		bind(DocumentListService.class).to(DocumentListServiceImpl.class).in(Scopes.SINGLETON);
 		bind(MailMetaDataConverter.class).in(Scopes.SINGLETON);
 		bind(DocumentHelper.class).in(Scopes.SINGLETON);
 		bind(LogActivities.class).in(Scopes.SINGLETON);
